requirements.txt
flask
llama-cpp-python
jinja2